remove useTournamentPods hook
