+++ client/src/views/MyMatchesView.tsx
@@ -11,7 +11,7 @@ import {
 import Typography from '@material-ui/core/Typography'
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
 import { useMatchesForUser } from '../hooks/useMatchesForUser'
-import { Match } from '../hooks/useTournamentPods'
+import { Match } from '../hooks/useTournamentPod'
 import { UserContext } from '../App'
 import { TournamentMatchView } from '../components/TournamentMatchView'
